+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - **[Fix]** Ping packets were not being sent on TCP connections during long transactions, causing premature timeouts
 - **[Fix]** Make sure ExceptionListeners are not called by mission-critical threads that cannot afford to wait
 - **[Feature]** Add more consistency checks during templates and destinations definitions load
+- **[Fix]** LocalQueueMBean getPersistentStoreUsage() should use absolute storage usage instead of current store usage if queue can auto-extend
 
  `FFMQ 4.0.x branch was forked from FFMQ 3.0.5`